End-to-end BC tests : moving helper functions and classes to utils.h/.cpp (part 1)
